requirements.txt
Flask
python-dotenv
requests
Flask-Limiter
Flask-Cors